csv2kml: replace model table with header parser
The field ordering can vary depending on the version of the tool
used to extract CSV data from black box files.

Instead of having a fixed mapping from model names to field sets,
generate a mapping on-the-fly by parsing the header row of the
file, and add a pair of options (-f/--field-map, -F/--field-file)
to allow specifying a field mapping on the command line, or via a
text file.

The -f/--file-map syntax is:

  "FIELD1:index1,FIELD2:index2,FIELD3:index3,...,FIELDN:indexN"

The -F/--field-file syntax uses one field per line with each
line formatted as "FIELDN:indexN":

  FIELD1:index1
  FIELD2:index2
  ...